Task 3.5  Create pipes
