+++ courses-app/src/app/shared/components/course-card/course-card.component.ts
@@ -18,23 +18,6 @@ export class CourseCardComponent {
   } = {id: '', title: '', description: '', creationDate: 'new Date()', duration: 0,  authors: [], isEditable: false}
   @Input() isCourseEditable = false;
 
-  getCourseDurationText () {
-    const minutes = this.courseInfo.duration % 60
-    const hours = (this.courseInfo.duration - minutes) / 60
-    const courseDurationText = `${withLeadZero(String(hours))}:${withLeadZero(String(minutes))}`
-
-    return courseDurationText
-  }
-
-  getCourseCreationDateText () {
-    const creationDateAsDate = new Date(this.courseInfo.creationDate)
-    const day = creationDateAsDate.getDate()
-    const month = creationDateAsDate.getMonth() + 1
-    const year = creationDateAsDate.getFullYear()
-
-    return `${withLeadZero(String(day))}.${withLeadZero(String(month))}.${year}`
-  }
-
   getAuthorsText () {
     return this.courseInfo.authors.join(', ')
   }
